editoast: move ApplicableDirectionsTrackRange to editoast_schemas
