+++ editoast/editoast_schemas/src/infra.rs
@@ -1,4 +1,5 @@
 mod applicable_directions;
+mod applicable_directions_track_range;
 mod direction;
 mod directional_track_range;
 mod endpoint;
@@ -11,6 +12,7 @@ mod track_range;
 mod waypoint;
 
 pub use applicable_directions::ApplicableDirections;
+pub use applicable_directions_track_range::ApplicableDirectionsTrackRange;
 pub use direction::Direction;
 pub use directional_track_range::DirectionalTrackRange;
 pub use endpoint::Endpoint;
