+++ editoast/src/converters/utils.rs
@@ -1,4 +1,5 @@
 use editoast_schemas::infra::ApplicableDirections;
+use editoast_schemas::infra::ApplicableDirectionsTrackRange;
 use editoast_schemas::infra::Direction;
 use std::collections::HashMap;
 use std::str::FromStr;
