+++ editoast/src/schema/mod.rs
@@ -19,7 +19,6 @@ pub use buffer_stop::BufferStop;
 pub use buffer_stop::BufferStopCache;
 pub use detector::Detector;
 pub use detector::DetectorCache;
-use editoast_schemas::infra::ApplicableDirections;
 use editoast_schemas::primitives::OSRDIdentified;
 use editoast_schemas::primitives::ObjectType;
 pub use electrification::Electrification;
@@ -63,42 +62,9 @@ cfg_if! {
 }
 
 use crate::infra_cache::operation;
-use derivative::Derivative;
-use serde::Deserialize;
-use serde::Serialize;
-
-use editoast_common::Identifier;
 
 editoast_common::schemas! {
     utils::schemas(),
     editoast_schemas::schemas(),
     operation::schemas(),
-}
-
-#[derive(Debug, Derivative, Clone, Deserialize, Serialize, PartialEq)]
-#[serde(deny_unknown_fields)]
-#[derivative(Default)]
-pub struct ApplicableDirectionsTrackRange {
-    #[derivative(Default(value = r#""InvalidRef".into()"#))]
-    pub track: Identifier,
-    pub begin: f64,
-    #[derivative(Default(value = "100."))]
-    pub end: f64,
-    pub applicable_directions: ApplicableDirections,
-}
-
-impl ApplicableDirectionsTrackRange {
-    pub fn new<T: AsRef<str>>(
-        track: T,
-        begin: f64,
-        end: f64,
-        applicable_directions: ApplicableDirections,
-    ) -> Self {
-        Self {
-            track: track.as_ref().into(),
-            begin,
-            end,
-            applicable_directions,
-        }
-    }
 }
