+++ editoast/src/views/pathfinding/electrifications.rs
@@ -150,10 +150,10 @@ pub mod tests {
     use crate::models::pathfinding::tests::simple_pathfinding_fixture;
     use crate::modelsv2::prelude::*;
     use crate::modelsv2::ElectrificationModel;
-    use crate::schema::ApplicableDirectionsTrackRange;
     use crate::schema::Electrification as ElectrificationSchema;
     use crate::views::tests::create_test_service;
     use editoast_schemas::infra::ApplicableDirections;
+    use editoast_schemas::infra::ApplicableDirectionsTrackRange;
 
     #[fixture]
     fn simple_mode_map() -> TrackMap<String> {
